add printf / sprintf routines

+++ kernel/sched.c
@@ -18,6 +18,7 @@
 #include "sched.h"
 
 #include "processor.h"
+#include "stddef.h"
 
 #define MAX_PRIO  3
 #define IDLE_PRIO 0
@@ -34,8 +35,8 @@ extern void _switch_to(thread_t *prev_thread, thread_t *next_thread);
  ******************************************************************************/
 extern stack_t   idle_stack;
 task_t idle_task __idle_task_data = {
-    .task_id.vms_id    = NULL,
-    .task_id.thread_id = NULL,
+    .task_id.vms_id    = (uint64_t)NULL,
+    .task_id.thread_id = (uint64_t)NULL,
     .prio              = IDLE_PRIO,
     .state             = READY,
     .stack             = &idle_stack,
